CI: skip regular `go test` on linux
See the added comment.
An alternative would be to split up the work into more jobs,
but that doesn't feel necessary, and we still have a global limit
on how many free runners we can run jobs on at once.

While here, update the comments, and remove unnecessary step names.