replace listedPackage.Deps in favor of a derived map
In looking at the cpu and memory profiles, it surfaced that we spent
a lot of time in garbage collection, and a significant amount of the
garbage was produced by gob decoding string slices.

listedPackage.Deps is a list of a package's transitive dependencies,
so as a Go build gets larger, the list also gets larger and larger.

Given that Imports is the list of direct dependencies,
we can reconstruct it ourselves as needed, which is not always.
Moreover, since we want to do lookups, we can build a map directly.

This doesn't directly result in a wall time speed-up,
but it does result in a significant reduction in allocations.
The gob files we store in the disk cache should also be a bit smaller.

            │      old      │               new               │
            │ cached-sec/op │ cached-sec/op  vs base          │
    Build-8     339.5m ± 2%     340.3m ± 1%  ~ (p=0.218 n=10)

            │     old     │                new                 │
            │ mallocs/op  │ mallocs/op   vs base               │
    Build-8   38.08M ± 0%   35.73M ± 0%  -6.18% (p=0.000 n=10)