store name pairs for _realName as a slice of pairs
Iterating over a map is much more expensive than iterating over a slice,
given how it needs to work out which keys are present in each bucket
and then randomize the order in which to navigate the keys.
None of this work needs to happen when iterating over a slice.

A map would be nice if we were to actually do map lookups, but we don't.

                  │     old     │                new                 │
                  │   sec/op    │   sec/op     vs base               │
    AbiRealName-8   707.1µ ± 1%   196.7µ ± 1%  -72.17% (p=0.001 n=7)

                  │     old      │                  new                  │
                  │     B/s      │      B/s       vs base                │
    AbiRealName-8   517.6Ki ± 2%   1816.4Ki ± 1%  +250.94% (p=0.001 n=7)

                  │     old      │                new                 │
                  │     B/op     │     B/op      vs base              │
    AbiRealName-8   5.362Ki ± 0%   5.359Ki ± 0%  -0.05% (p=0.001 n=7)

                  │    old     │              new              │
                  │ allocs/op  │ allocs/op   vs base           │
    AbiRealName-8   19.00 ± 0%   19.00 ± 0%  ~ (p=1.000 n=7) ¹