clarify internal/abi name code a bit
Use the term "original name" rather than "real name" for the code
as it is clearer that we mean the unobfuscated name.

Update the comments at the top of the file to be clearer with the
explanation of what kinds of inputs we can expect.

While here, use fmt.Appendf to simplify the generation code a bit.