expose cgo issue 916
Seems to happen when the main package only has Go files importing "C",
meaning that it has zero "pure Go" files.

To avoid needing two main package Go builds for cgo.txtar,
switch our main package to this scenario as it seems more interesting.

While here, add a test case for a Go callback function taking a C param
as that is relatively common and we had no coverage for it.

This only reproduces the bug; the fix is coming separately.

For #916.